Split ingestion controllers and move controller tests into matching packages.
