+++ app/controllers/ingestion/CustomsOfficeListController.scala
@@ -32,44 +32,17 @@ import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
 
-class ReferenceDataController @Inject() (
+class CustomsOfficeListController @Inject() (
   cc: ControllerComponents,
   referenceDataService: ReferenceDataService,
   parseConfig: ReferenceDataControllerParserConfig,
-  cTCUP06Schema: CTCUP06Schema,
   cTCUP08Schema: CTCUP08Schema
 )(implicit ec: ExecutionContext)
     extends BackendController(cc) {
 
   import parseConfig._
 
-  private val referenceDataListsLogger = Logger("ReferenceDataLists")
   private val customsOfficeListsLogger = Logger("CustomsOfficeLists")
-
-  def referenceDataLists(): Action[RawBuffer] =
-    Action(referenceDataParser(parse)).async {
-
-      implicit request =>
-        val requestBody = request.body.asBytes().map(_.toArray) match {
-          case Some(body) => referenceDataService.validateAndDecompress(cTCUP06Schema, body)
-          case _          => Left(OtherError("Payload larger than memory threshold"))
-        }
-
-        requestBody match {
-          case Right(jsObject) =>
-            referenceDataService
-              .insert(ReferenceDataListsPayload(jsObject))
-              .map {
-                case DataProcessingSuccessful => Accepted
-                case DataProcessingFailed =>
-                  referenceDataListsLogger.error("Failed to save the data list because of internal error")
-                  InternalServerError(Json.toJsObject(OtherError("Failed in processing the data list")))
-              }
-          case Left(error) =>
-            referenceDataListsLogger.error(Json.toJsObject(error).toString())
-            Future.successful(BadRequest(Json.toJsObject(error)))
-        }
-    }
 
   def customsOfficeLists(): Action[RawBuffer] =
     Action(customsOfficeParser(parse)).async {
