+++ conf/app.routes
@@ -3,5 +3,5 @@
 GET     /customs-reference-data/lists                   controllers.consumption.ResourceLinksController.get
 GET     /customs-reference-data/lists/:listName         controllers.consumption.ListRetrievalController.get(listName: ListName)
 
-POST    /reference-data-lists                           controllers.ingestion.ReferenceDataController.referenceDataLists()
-POST    /customs-office-lists/customs-office-lists      controllers.ingestion.ReferenceDataController.customsOfficeLists
+POST    /reference-data-lists                           controllers.ingestion.ReferenceDataListController.referenceDataLists()
+POST    /customs-office-lists/customs-office-lists      controllers.ingestion.CustomsOfficeListController.customsOfficeLists()
